+++ doc/changelog.rst
@@ -31,6 +31,8 @@ Bug fixes
 ^^^^^^^^^
 - Fixed a bug where ``actuator_force`` was not set in MJX (:github:issue:`2068`).
 - Fixed bug where MJX data tendon fields were incorrect after calling ``mjx.put_data``.
+- The compiler now returns an error if height fields are used with :ref:`collision sensors<collision-sensors>` as they
+  are not yet supported.
 
 
 Version 3.2.3 (Sep 16, 2024)